+++ addon/button.js
@@ -2,6 +2,12 @@
 /* global showStdPageDetails */
 
 
+(function() {
+  let script = document.createElement('script');
+  script.src = chrome.runtime.getURL('inject.js');
+  document.body.appendChild(script);
+})();
+
 // sfdcBody = normal Salesforce page
 // ApexCSIPage = Developer Console
 // auraLoadingBox = Lightning / Salesforce1
@@ -194,6 +200,9 @@ function initButton(sfHost, inInspector) {
         } else {
           document.querySelectorAll("." + apiNamesClass).forEach(e => e.remove());
         }
+      }
+      if (e.data.lightningNavigate) {
+        document.dispatchEvent(new CustomEvent("lightningNavigate", {detail: e.data.lightningNavigate}));
       }
     });
     rootEl.appendChild(popupEl);